neo4j 初步开始

+++ QueryServer/src/main/java/com/f4/DWQueryServer/entity/answer/DataAnswer.java
@@ -13,5 +13,5 @@ import java.util.List;
 @Data
 public class DataAnswer {
     private List<String> data;
-    private Integer time;
+    private Long time;
 }
